feat: 流水线插件支持展示获得的荣誉和SLA等信息 #8114

+++ src/backend/ci/core/metrics/api-metrics/src/main/kotlin/com/tencent/devops/metrics/pojo/po/SaveAtomIndexStatisticsDailyPO.kt
@@ -38,9 +38,9 @@ data class SaveAtomIndexStatisticsDailyPO(
     @ApiModelProperty("插件代码")
     val atomCode: String,
     @ApiModelProperty("失败执行次数")
-    var failExecuteCount: Long,
+    var failExecuteCount: Int,
     @ApiModelProperty("失败合规次数")
-    var failComplianceCount: Long,
+    var failComplianceCount: Int,
     @ApiModelProperty("统计时间")
     val statisticsTime: LocalDateTime,
     @ApiModelProperty("创建人")
